+++ DesktopReact/frontend/src/api/follow.httpService.js
@@ -1,0 +1,19 @@
+import {
+    baseURL,
+    GET, 
+    POST,
+    DELETE
+} from './httpConfig'
+
+const followUser = (userProfileId) =>{
+   return httpRequest('/follow', POST, { userProfileId: userProfileId })
+}
+
+function unfollowUser(userProfileId) {
+    return httpRequest('/unFollow', POST, { userProfileId: userProfileId })
+}
+
+export {
+    followUser,
+    unfollowUser
+}
