+++ DesktopReact/frontend/src/api/post.httpService.js
@@ -1,0 +1,18 @@
+import {
+    httpRequest,
+    GET,
+    POST
+} from './httpConfig'
+
+function createPost(postBody) {
+    return httpRequest('/createPost', POST, postBody);
+}
+
+function getPosts(postBody) {
+    return httpRequest('/getPosts', GET, postBody);
+}
+
+export {
+    createPost, 
+    getPosts
+}
