+++ DesktopReact/frontend/src/api/profile.httpService.js
@@ -1,0 +1,13 @@
+import {
+    GET,
+    httpRequest
+} from './httpConfig'
+
+function getProfiles() {
+    return httpRequest('/profiles', GET);
+}
+
+
+export {
+    getProfiles
+}
