+++ DesktopReact/frontend/src/components/postPlaceholder/PostPlaceholder.js
@@ -2,7 +2,8 @@ import "./PostPlaceholder.css"
 import React, {useEffect, useState} from 'react';
 import {Alert, Avatar, Button, Card, Snackbar, Stack, TextField, Typography} from "@mui/material";
 import pic from "../../images/superalgos.png"
-import {createPost, STATUS_OK} from "../../api/httpService";
+import { STATUS_OK }  from "../../api/httpConfig";
+import { createPost } from "../../api/post.httpService"
 
 const PostPlaceholder = () => {
     const [postText, setPostText] = useState(undefined);
